feat(utils): moved findSelection into utils.js
