+++ src/storage-client.js
@@ -1,5 +1,6 @@
 import {parseQueryString} from 'aurelia-path';
 import extend from 'extend';
+import {findSelected} from './utils';
 
 const baseStorageKey = 'AureliaStorageClient';
 
@@ -128,23 +129,4 @@ export class StorageClient {
       }
     }
   }
-}
-
-function findSelected(resource, queryParameters) {
-  if (Object.keys(queryParameters).length === 0) return resource;
-
-  let selection = [];
-
-  resource.map(el => {
-    let matches = true;
-    for (let key in queryParameters) {
-      if (el[key] === undefined || queryParameters[key] !== el[key].toString()) {
-        matches = false;
-        break;
-      }
-    }
-    if (matches) selection.push(el);
-  });
-
-  return selection;
 }
